requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
dblpy==0.4.0
discord-buttons-plugin==0.1.2
DiscordUtils==1.3.4
Django==3.2.7
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
humanize==3.11.0
hyperframe==5.2.0
idna==2.10
ksoftapi==0.4.1
motor==2.5.1
multidict==5.1.0
Pillow==8.3.1
py-cord==1.7.3
pycparser==2.20
pymongo==3.12.0
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
randfacts==0.9.0
ratelimiter==1.2.0.post0
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
spotipy==2.19.0
sqlparse==0.4.1
timeago==1.0.15
typing-extensions==3.10.0.0
urllib3==1.26.6
utils==1.0.1
wavelink==0.9.10
yarl==1.6.3
youtube-dl==2021.6.6
